+++ packages/web3-errors/src/errors/transaction_errors.ts
@@ -26,6 +26,8 @@ import {
 	ERR_TX_CHAIN_ID_MISMATCH,
 	ERR_TX_DATA_AND_INPUT,
 	ERR_TX_GAS_MISMATCH,
+	ERR_TX_CHAIN_MISMATCH,
+	ERR_TX_HARDFORK_MISMATCH,
 	ERR_TX_INVALID_CALL,
 	ERR_TX_INVALID_CHAIN_INFO,
 	ERR_TX_INVALID_FEE_MARKET_GAS,
@@ -206,6 +208,22 @@ export class ChainIdMismatchError extends InvalidValueError {
 			// https://github.com/ChainSafe/web3.js/blob/8783f4d64e424456bdc53b34ef1142d0a7cee4d7/packages/web3-eth-accounts/src/index.js#L176
 			'Chain Id doesnt match in tx.chainId tx.common.customChain.chainId',
 		);
+	}
+}
+
+export class ChainMismatchError extends InvalidValueError {
+	public code = ERR_TX_CHAIN_MISMATCH;
+
+	public constructor(value: { txChain: unknown; baseChain: unknown }) {
+		super(JSON.stringify(value), 'Chain doesnt match in tx.chain tx.common.basechain');
+	}
+}
+
+export class HardforkMismatchError extends InvalidValueError {
+	public code = ERR_TX_HARDFORK_MISMATCH;
+
+	public constructor(value: { txHardfork: unknown; commonHardfork: unknown }) {
+		super(JSON.stringify(value), 'hardfork doesnt match in tx.hardfork tx.common.hardfork');
 	}
 }