+++ packages/web3-validator/test/unit/default_validator.test.ts
@@ -15,14 +15,12 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-export const namehashValidData: [string, string][] = [
-	['', '0x0000000000000000000000000000000000000000000000000000000000000000'],
-	['eth', '0x93cdeb708b7545dc668eb9280176169d1c33cfd8ed6f04690a0bcc88a93fc4ae'],
-	['foo.eth', '0xde9b09fd7c5f901e23a3f19fecc54828e9c848539801e86591bd9801b019f84f'],
-];
+import { validator } from '../../src/default_validator';
+import { Web3Validator } from '../../src/web3_validator';
 
-export const normalizeValidData: [string, string][] = [
-	['öbb.at', 'xn--bb-eka.at'],
-	['xn--bb-eka.at', 'xn--bb-eka.at'],
-	['XN--BB-EKA.AT', 'xn--bb-eka.at'],
-];
+describe('default web3-validator', () => {
+	it('should be instance', () => {
+		expect(validator).toBeDefined();
+		expect(validator).toBeInstanceOf(Web3Validator);
+	});
+});