+++ packages/web3-eth/CHANGELOG.md
@@ -40,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   `web3-rpc-methods` dependency (#5441)
+-   added chain and hardfork validation for transaction and transaction.common object in `validateTransactionForSigning`
 
 ### Changed